backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
pyarrow==14.0.1
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
implicit==0.7.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
lightgbm==4.1.0
redis==5.0.1
structlog==23.2.0
prometheus-client==0.19.0
pydantic==2.5.0
pytest==7.4.3
python-multipart==0.0.6